external/chre: Fix sign-compare warnings in util/tests
e.g.
```
/PATH/TO/REPO/apps/external/googletest/googletest/googletest/include/gtest/gtest.h: In instantiation of 'testing::AssertionResult testing::internal::CmpHelperEQ(const char*, const char*, const T1&, const T2&) [with T1 = unsigned int; T2 = int]':
/PATH/TO/REPO/apps/external/googletest/googletest/googletest/include/gtest/gtest.h:1400:23:   required from 'static testing::AssertionResult testing::internal::EqHelper::Compare(const char*, const char*, const T1&, const T2&) [with T1 = unsigned int; T2 = int; typename std::__1::enable_if<((! std::__1::is_integral<_Tp>::value) || (! std::__1::is_pointer<_Deleter>::value))>::type* <anonymous> = 0]'
/PATH/TO/CHRE/util/tests/buffer_test.cc:33:3:   required from here
/PATH/TO/REPO/apps/external/googletest/googletest/googletest/include/gtest/gtest.h:1381:11: warning: comparison of integer expressions of different signedness: 'const unsigned int' and 'const int' [-Wsign-compare]
 1381 |   if (lhs == rhs) {
      |       ~~~~^~~~~~
```

+++ util/tests/debug_dump_test.cc
@@ -35,7 +35,7 @@ TEST(DebugDumpWrapper, OneBufferForOneString) {
   const char *str = "Lorem ipsum";
   debugDump.print("%s", str);
   const auto &buffers = debugDump.getBuffers();
-  EXPECT_EQ(buffers.size(), 1);
+  EXPECT_EQ(buffers.size(), (size_t)1);
   EXPECT_TRUE(strcmp(buffers.front().get(), str) == 0);
 }
 
@@ -46,7 +46,7 @@ TEST(DebugDumpWrapper, TwoStringsFitPerfectlyInOneBuffer) {
   debugDump.print("%s", str1);
   debugDump.print("%s", str2);
   const auto &buffers = debugDump.getBuffers();
-  EXPECT_EQ(buffers.size(), 1);
+  EXPECT_EQ(buffers.size(), (size_t)1);
   char bothStr[5];
   strcpy(bothStr, str1);
   strcat(bothStr, str2);
@@ -60,12 +60,12 @@ TEST(DebugDumpWrapper, TooLargeOfStringToFit) {
   const auto &buffers = debugDump.getBuffers();
 
   // One null-terminated buffer will be created for an empty wrapper.
-  EXPECT_EQ(buffers.size(), 1);
+  EXPECT_EQ(buffers.size(), (size_t)1);
   EXPECT_TRUE(strcmp(buffers.back().get(), "") == 0);
 
   // Once there's a buffer, it won't be updated.
   debugDump.print("%s", str);
-  EXPECT_EQ(buffers.size(), 1);
+  EXPECT_EQ(buffers.size(), (size_t)1);
   EXPECT_TRUE(strcmp(buffers.back().get(), "") == 0);
 }
 
@@ -75,9 +75,9 @@ TEST(DebugDumpWrapper, TooLargeOfStringWithPartlyFilledBuffer) {
   const char *str2 = "bc";
   debugDump.print("%s", str1);
   const auto &buffers = debugDump.getBuffers();
-  EXPECT_EQ(buffers.size(), 1);
+  EXPECT_EQ(buffers.size(), (size_t)1);
   debugDump.print("%s", str2);
-  EXPECT_EQ(buffers.size(), 1);
+  EXPECT_EQ(buffers.size(), (size_t)1);
   EXPECT_TRUE(strcmp(buffers.front().get(), str1) == 0);
 }
 
@@ -88,7 +88,7 @@ TEST(DebugDumpWrapper, StringForcesNewBufferWithPartlyFilledBuffer) {
   debugDump.print("%s", str1);
   debugDump.print("%s", str2);
   const auto &buffers = debugDump.getBuffers();
-  EXPECT_EQ(buffers.size(), 2);
+  EXPECT_EQ(buffers.size(), (size_t)2);
   EXPECT_TRUE(strcmp(buffers.front().get(), str1) == 0);
   EXPECT_TRUE(strcmp(buffers.back().get(), str2) == 0);
 }
@@ -102,14 +102,14 @@ TEST(DebugDumpWrapper, ManyNewBuffersAllocated) {
     debugDump.print("%s", str);
   }
   const auto &buffers = debugDump.getBuffers();
-  EXPECT_EQ(buffers.size(), 3);
+  EXPECT_EQ(buffers.size(), (size_t)3);
 }
 
 TEST(DebugDumpWrapper, EmptyStringAllocsOneBuffer) {
   DebugDumpWrapper debugDump(kStandardBufferSize);
   debugDump.print("%s", "");
   const auto &buffers = debugDump.getBuffers();
-  EXPECT_EQ(buffers.size(), 1);
+  EXPECT_EQ(buffers.size(), (size_t)1);
 }
 
 TEST(DebugDumpWrapper, BuffersClear) {
@@ -121,15 +121,15 @@ TEST(DebugDumpWrapper, BuffersClear) {
   debugDump.print("%s", str1);
   debugDump.print("%s", str2);
   const auto &buffers = debugDump.getBuffers();
-  EXPECT_EQ(buffers.size(), 2);
+  EXPECT_EQ(buffers.size(), (size_t)2);
   EXPECT_TRUE(strcmp(buffers.front().get(), str1) == 0);
   EXPECT_TRUE(strcmp(buffers.back().get(), str2) == 0);
 
   debugDump.clear();
-  EXPECT_EQ(buffers.size(), 0);
+  EXPECT_EQ(buffers.size(), (size_t)0);
 
   debugDump.print("%s", str3);
-  EXPECT_EQ(buffers.size(), 1);
+  EXPECT_EQ(buffers.size(), (size_t)1);
   EXPECT_TRUE(strcmp(buffers.front().get(), str3) == 0);
 }
 
@@ -147,7 +147,7 @@ TEST(DebugDumpWrapper, PrintVaListTwoStrings) {
   printVaList(&debugDump, "%s", str1);
   printVaList(&debugDump, "%s", str2);
   const auto &buffers = debugDump.getBuffers();
-  EXPECT_EQ(buffers.size(), 1);
+  EXPECT_EQ(buffers.size(), (size_t)1);
   char bothStr[5];
   strcpy(bothStr, str1);
   strcat(bothStr, str2);
